+++ elmyra.ip.access.epo/CHANGES.rst
@@ -11,6 +11,7 @@ development
 - middleware: add some request variables (host-, and path-information) to application configuration
 - ui: make permalink buttons show popovers with full uris in text input form elements
 - ui, storage: resolve “duplicate project name problem” when importing database to "context=viewer"
+- ui: improve permalink mechanics
 
 0.22.2
 ======